Class inheritance for animals, and overwrite method
