+++ exercises/petting_zoo/models/animals/animals.py
@@ -1,3 +1,22 @@
+from datetime import date
+class Animal:
+    def __init__(self, name, species, food, chip_num):
+        self.name = name
+        self.species = species
+        self.food = food
+        self.__chip_number = chip_num
+        self.date_added = date.today()
+
+    def feed(self):
+        print(f'{self.name} was fed {self.food} on {date.today().strftime("%m/%d/%Y")}')
+    
+    @property
+    def chip_number(self):
+        return self.__chip_number
+    
+    @chip_number.setter
+    def chip_number(self, num):
+        pass
 
 
 
@@ -5,17 +24,3 @@
 
 
 
-
-
-# glenn = Goat('Glenn', 'nigerian')
-# kuzco = Llama('Kuzco', 'demestic')
-# bubbles = Fish('Bubbles', 'parana')
-# fang = Snake('Fang', ' Corn snake' )
-# penelope = Pig('Penelope', 'Popbellied')
-# bunkie = Cat('Bunkie', 'tabby')
-# rough = Dog('Rough', 'Lab')
-# shy = Sheep()
-# nah = Horse()
-# leo = Lizard()
-
-# print(glenn, kuzco, bubbles, fang, penelope, bunkie, rough )
