+++ exercises/petting_zoo/models/animals/goat.py
@@ -1,24 +1,8 @@
 from datetime import date
+from .animals import Animal
 
-class Goat():
-    def __init__(self, name, species, shift, food, chip_number):
-        self.name = name
-        self.species = species
-        self.date_added = date.today()
+class Goat(Animal):
+    def __init__(self, name, species, shift, food, chip_num):
+        super().__init__(name, species,food, chip_num)
         self.walking = True
         self.shift = shift
-        self.food = food
-        self.chip_number = chip_number
-
-    def feed(self):
-        print(f'{self.name} ate {self.food} at {date.today().strftime("%m/%d/%Y")}')
-    
-    @property
-    def chip_number(self):
-        return self.__chip_number # Note there are 2 underscores here
-      
-
-
-    @chip_number.setter
-    def chip_number(self, number):
-        pass
